Move loss methods to separate module
