+++ configs/loss/semantic_loss.yml
@@ -1,4 +1,4 @@
-class_path: chebai.models.electra.ElectraChEBIDisjointLoss
+class_path: chebai.loss.semantic.DisjointLoss
 init_args:
   path_to_chebi: data/ChEBI100/raw/chebi.obo
   path_to_label_names: data/ChEBI100/raw/classes.txt
